Christoph Junghans reported in GenToo #424137 a build problem which was reproduced with this example: 	mawk 'BEGIN {getline <"/dev/stdin" }' which I tested as 	mawk 'BEGIN {getline <"/dev/stdin"; print $1; }'
--------Analysis
He asserted that it worked with an earlier version (than p20120627).

p20120627 introduced this change:
	+ add "/dev/stdin" name for stdin for consistency with other
	  implementations of awk (suggested by Aharon Robbins).

Before that change mawk did not treat /dev/stdin specially (and should not
have worked at all).  Testing shows it does -- on Linux, which has that
special device.  I overlooked that, since I recall it as nonportable, and
don't use nonportable features.

Checking now, these support the feature
	Solaris 6-10
	BSD's (FreeBSD, NetBSD, OpenBSD)
	Linux
	Cygwin

These do not
	IRIX64
	HPUX 10.20 - 11.23 - 11.31
	Tru64 5.1
	AIX 5.1 - 7.1
	Win32 (ignoring msys shell)

--------Fix
The bug was introduced by assigning "stdin" to the "ptr" member in set_stdio().
It turns out that file_find() uses a FILE* pointer for inputs, and a FIN*
pointer for input.  Then (because open("/dev/stdin") "just works" on Linux)
the FINopen() call seemed to work before.  Setting the ptr member stopped
file_find() from calling FINopen(), and returned FILE* data as FIN* data.

For portability, I want to call FINopen() and have it make provision for
"/dev/stdin" in the same way that "-" was used as a special alias for it.
To do this, file_find() has to be modified to handle /dev/stdin as a special
case.  Modify set_stdio() to use a null pointer in the table entry to get
past the table lookup in file_find(), and add a check for a null "ptr" as
input - and _then_ call FINopen().

The logic flow in file_find() couldn't handle this change.  I ironed it out
(getting rid of goto's), and now it's more readable as well.